[Feature] Moved the onboarding steps to a subfolder
